techweekia9-multi-agents-rag-main/services/agents/pyproject.toml
[project]
name = "legal-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk==1.27.0",
    "litellm>=1.50.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agents", "tools", "ingestion"]
